Add `Mirage_block.Make_safe_BLOCK` functor
This functor wraps `read` and `write` to check preconditions of the
passed buffers and to fail with a useful error message if a precondition
doesn't hold. Currently the only precondition is that buffers must be
a multiple of sector_size in length.

Example usage:
```
  Ramdisk.connect "foo"
  >>= fun x ->
  let module Safe = Mirage_block.Make_safe_BLOCK(Ramdisk) in
  Safe.read x 0L [ buf ]
  ...
```

Note this patch also adds a dependency on `logs` and will emit a log
message if it encounters a bad buffer.

Signed-off-by: David Scott <dave.scott@docker.com>